Refactor library module and increment version to 2.0
In order for this to become a comprehensive view library (as in #30), the package name needed to be changed. Therefore files from the library module have been moved so that the package name is `com.satsuware.usefulviews`.

The change from `.labelledspinner` to `.usefulviews` is because of the reason stated above, however the change from `com.farbod` to `com.satsuware` is due to the fact that I did not own the domain 'farbod.com', so it would be better to use a domain that I did own.
The domain 'satsuware.com' (also (and short for) 'satsumasoftware.com') will be used instead. Furthermore, the plan will be to upload documentation for this project on this website (see #24).

In addition to these changes, the `build.gradle` file has been changed so that the library name and description correspond with the idea of a library for a collection of views as opposed to just having the `LabelledSpinner`.

To signify these changes, the version has been incremented from 1.2 to 2.0.