Refactor sample module
As I wrote in commit 5970e68 :

In order for this to become a comprehensive view library (as in #30), the package name needed to be changed. Therefore files from the sample module have been moved so that the package name is `com.satsuware.usefulviews`.

The change from `.labelledspinner` to `.usefulviews` is because of the reason stated above, however the change from `com.farbod` to `com.satsuware` is due to the fact that I did not own the domain 'farbod.com', so it would be better to use a domain that I did own.
The domain 'satsuware.com' (also (and short for) 'satsumasoftware.com') will be used instead. Furthermore, the plan will be to upload documentation for this project on this website (see #24).

To comply with these changes the `build.gradle` file has been modified to compile with the new dependency.