Fix flip issue and add option to remove views
One issue that was being caused was that the widget could not be flipped if it had not been turned the other way when the Views were being switched. This has been fixed by reassigning a boolean variable when it updates.

The option to remove the front and back Views from the widget has also been added.